requirements.txt
absl-py==1.4.0
accelerate==0.29.2
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
augly==1.0.0
braceexpand==0.1.7
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.1
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.4
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.11.0
datasets==2.13.1
dill==0.3.6
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
faiss-cpu==1.7.4
faiss-gpu==1.7.2
filelock @ file:///croot/filelock_1672387128942/work
fonttools==4.40.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
huggingface-hub==0.22.2
idna @ file:///croot/idna_1666125576474/work
imageio==2.32.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
invisible-watermark==0.2.0
iopath==0.1.10
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
llvmlite==0.41.1
Markdown==3.4.3
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcls==0.23.2
mmcv-full==1.6.1
mmsegmentation==0.27.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
multiprocess==0.70.14
networkx @ file:///croot/networkx_1690561992265/work
nltk==3.8.1
numba==0.58.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.1.78
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
peft==0.8.2
Pillow @ file:///croot/pillow_1696580024257/work
pillow_heif==0.16.0
platformdirs==3.8.1
portalocker==2.7.0
prettytable==3.4.1
promise==2.3
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.4.2
pydantic_core==2.10.1
pynndescent==0.5.13
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-magic==0.4.27
pytorchcv==0.0.67
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2023.6.3
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.28.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sympy @ file:///croot/sympy_1668202399572/work
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.2
tokenizers==0.15.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.39.3
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
umap-learn==0.5.6
urllib3 @ file:///croot/urllib3_1698257533958/work
wandb==0.15.5
wcwidth==0.2.9
webdataset==0.2.48
Werkzeug==2.3.6
wordcloud==1.9.3
xxhash==3.2.0
yapf==0.40.1
yarl==1.9.2
zipp==3.16.0
